Add an mismatch check to context saves
