+++ packages/docregistry/src/context.ts
@@ -200,7 +200,6 @@ class Context<T extends DocumentRegistry.IModel> implements DocumentRegistry.ICo
    */
   save(): Promise<void> {
     let model = this._model;
-    let path = this._path;
     if (model.readOnly) {
       return Promise.reject(new Error('Read only'));
     }
@@ -218,7 +217,10 @@ class Context<T extends DocumentRegistry.IModel> implements DocumentRegistry.ICo
     };
 
     return this._manager.ready.then(() => {
-      return this._manager.contents.save(path, options);
+      if (!model.modelDB.isCollaborative) {
+        return this._maybeSave(options);
+      }
+      return this._manager.contents.save(this._path, options);
     }).then(value => {
       if (this.isDisposed) {
         return;
@@ -451,6 +453,57 @@ class Context<T extends DocumentRegistry.IModel> implements DocumentRegistry.ICo
     }).then(() => {
       this._isReady = true;
       this._populatedPromise.resolve(void 0);
+    });
+  }
+
+  /**
+   * Save a file, dealing with conflicts.
+   */
+  private _maybeSave(options: Partial<Contents.IModel>): Promise<Contents.IModel> {
+    let path = this._path;
+    // Make sure the file has not changed on disk.
+    let promise = this._manager.contents.get(path);
+    return promise.then(model => {
+      if (this.isDisposed) {
+        return Promise.reject('Disposed');
+      }
+      let modified = this.contentsModel && this.contentsModel.last_modified;
+      if (model.last_modified !== modified) {
+        return this._timeConflict(modified, model, options);
+      }
+      return this._manager.contents.save(path, options);
+    }, (err) => {
+      return this._manager.contents.save(path, options);
+    });
+  }
+
+  /**
+   * Handle a time conflict.
+   */
+  private _timeConflict(local: string, model: Contents.IModel, options: Partial<Contents.IModel>): Promise<Contents.IModel> {
+    let localTime = new Date(local);
+    let remoteTime = new Date(model.last_modified);
+    console.warn(`Last saving peformed ${localTime} ` +
+                 `while the current file seem to have been saved ` +
+                 `${remoteTime}`);
+    let body = `The file has changed on disk since the last time we ` +
+               `opened or saved it. ` +
+               `Do you want to overwrite the file on disk with the version ` +
+               ` open here, or load the version on disk (revert)?`;
+    let revertBtn = Dialog.okButton({ label: 'REVERT' });
+    let overwriteBtn = Dialog.warnButton({ label: 'OVERWRITE' });
+    return showDialog({
+      title: 'File Changed', body,
+      buttons: [Dialog.cancelButton(), revertBtn, overwriteBtn]
+    }).then(result => {
+      if (this.isDisposed) {
+        return Promise.reject('Disposed');
+      }
+      if (result.button.label === 'OVERWRITE') {
+        return this._manager.contents.save(this._path, options);
+      } else if (result.button.label === 'REVERT') {
+        return this.revert().then(() => { return model; });
+      }
     });
   }
